+++ packages/core/src/agent/agent.test.ts
@@ -3259,124 +3259,6 @@ function agentTests({ version }: { version: 'v1' | 'v2' }) {
       expect(Object.keys(tools).some(k => k.length === 63)).toBe(true);
       expect(Object.keys(tools)).not.toContain(longName);
     });
-
-    it('should make runtimeContext available to tools when injected in generate', async () => {
-      const testRuntimeContext = new RuntimeContext([['test-value', 'runtimeContext-value']]);
-      let capturedValue: string | null = null;
-
-      const testTool = createTool({
-        id: 'runtimeContext-test-tool',
-        description: 'A tool that verifies runtimeContext is available',
-        inputSchema: z.object({
-          query: z.string(),
-        }),
-        execute: ({ runtimeContext }) => {
-          capturedValue = runtimeContext.get('test-value')!;
-
-          return Promise.resolve({
-            success: true,
-            runtimeContextAvailable: !!runtimeContext,
-            runtimeContextValue: capturedValue,
-          });
-        },
-      });
-
-      const agent = new Agent({
-        name: 'runtimeContext-test-agent',
-        instructions: 'You are an agent that tests runtimeContext availability.',
-        model: openaiModel,
-        tools: { testTool },
-      });
-
-      const mastra = new Mastra({
-        agents: { agent },
-        logger: false,
-      });
-
-      const testAgent = mastra.getAgent('agent');
-
-      let response;
-      let toolCall;
-      if (version === 'v1') {
-        response = await testAgent.generateLegacy('Use the runtimeContext-test-tool with query "test"', {
-          toolChoice: 'required',
-          runtimeContext: testRuntimeContext,
-        });
-        toolCall = response.toolResults.find(result => result.toolName === 'testTool');
-      } else {
-        response = await testAgent.generate('Use the runtimeContext-test-tool with query "test"', {
-          toolChoice: 'required',
-          runtimeContext: testRuntimeContext,
-        });
-        toolCall = response.toolResults.find(result => result.payload.toolName === 'testTool').payload;
-      }
-
-      expect(toolCall?.result?.runtimeContextAvailable).toBe(true);
-      expect(toolCall?.result?.runtimeContextValue).toBe('runtimeContext-value');
-      expect(capturedValue).toBe('runtimeContext-value');
-    }, 500000);
-
-    it('should make runtimeContext available to tools when injected in stream', async () => {
-      const testRuntimeContext = new RuntimeContext([['test-value', 'runtimeContext-value']]);
-      let capturedValue: string | null = null;
-
-      const testTool = createTool({
-        id: 'runtimeContext-test-tool',
-        description: 'A tool that verifies runtimeContext is available',
-        inputSchema: z.object({
-          query: z.string(),
-        }),
-        execute: ({ runtimeContext }) => {
-          capturedValue = runtimeContext.get('test-value')!;
-
-          return Promise.resolve({
-            success: true,
-            runtimeContextAvailable: !!runtimeContext,
-            runtimeContextValue: capturedValue,
-          });
-        },
-      });
-
-      const agent = new Agent({
-        name: 'runtimeContext-test-agent',
-        instructions: 'You are an agent that tests runtimeContext availability.',
-        model: openaiModel,
-        tools: { testTool },
-      });
-
-      const mastra = new Mastra({
-        agents: { agent },
-        logger: false,
-      });
-
-      const testAgent = mastra.getAgent('agent');
-
-      let stream;
-      let toolCall;
-      if (version === 'v1') {
-        stream = await testAgent.streamLegacy('Use the runtimeContext-test-tool with query "test"', {
-          toolChoice: 'required',
-          runtimeContext: testRuntimeContext,
-        });
-
-        await stream.consumeStream();
-
-        toolCall = (await stream.toolResults).find(result => result.toolName === 'testTool');
-      } else {
-        stream = await testAgent.stream('Use the runtimeContext-test-tool with query "test"', {
-          toolChoice: 'required',
-          runtimeContext: testRuntimeContext,
-        });
-
-        await stream.consumeStream();
-
-        toolCall = (await stream.toolResults).find(result => result.payload.toolName === 'testTool').payload;
-      }
-
-      expect(toolCall?.result?.runtimeContextAvailable).toBe(true);
-      expect(toolCall?.result?.runtimeContextValue).toBe('runtimeContext-value');
-      expect(capturedValue).toBe('runtimeContext-value');
-    }, 500000);
   });
 
   describe(`${version} - context parameter handling`, () => {
